+++ Yunbeom/src/main/java/com/depromeet/yunbeom/common/controller/ExceptionControllerAdvice.java
@@ -1,0 +1,36 @@
+package com.depromeet.yunbeom.common.controller;
+
+import static org.springframework.http.HttpStatus.*;
+
+import org.springframework.core.Ordered;
+import org.springframework.core.annotation.Order;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+import com.depromeet.yunbeom.user.exception.CertificationCodeNotMatchedException;
+import com.depromeet.yunbeom.user.exception.ResourceNotFoundException;
+
+import lombok.RequiredArgsConstructor;
+
+@ControllerAdvice
+@Order(Ordered.HIGHEST_PRECEDENCE)
+@RequiredArgsConstructor
+public class ExceptionControllerAdvice {
+
+    @ResponseBody
+    @ResponseStatus(NOT_FOUND)
+    @ExceptionHandler(ResourceNotFoundException.class)
+    public String resourceNotFoundException(ResourceNotFoundException exception) {
+        return exception.getMessage();
+    }
+
+    @ResponseBody
+    @ResponseStatus(FORBIDDEN)
+    @ExceptionHandler(CertificationCodeNotMatchedException.class)
+    public String certificationCodeNotMatchedException(CertificationCodeNotMatchedException exception) {
+        return exception.getMessage();
+    }
+
+}
