--- Yunbeom/src/main/java/com/depromeet/yunbeom/controller/PostController.java
@@ -1,51 +1,0 @@
-package com.depromeet.yunbeom.controller;
-
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import com.depromeet.yunbeom.model.dto.PostResponse;
-import com.depromeet.yunbeom.model.dto.PostUpdateDto;
-import com.depromeet.yunbeom.repository.PostEntity;
-import com.depromeet.yunbeom.service.PostService;
-
-import io.swagger.v3.oas.annotations.tags.Tag;
-import lombok.RequiredArgsConstructor;
-
-@Tag(name = "게시물(posts)")
-@RestController
-@RequestMapping("/api/posts")
-@RequiredArgsConstructor
-public class PostController {
-
-    private final PostService postService;
-    private final UserController userController;
-
-    @GetMapping("/{id}")
-    public ResponseEntity<PostResponse> getPostById(@PathVariable long id) {
-        return ResponseEntity
-            .ok()
-            .body(toResponse(postService.getById(id)));
-    }
-
-    @PutMapping("/{id}")
-    public ResponseEntity<PostResponse> updatePost(@PathVariable long id, @RequestBody PostUpdateDto postUpdateDto) {
-        return ResponseEntity
-            .ok()
-            .body(toResponse(postService.update(id, postUpdateDto)));
-    }
-
-    public PostResponse toResponse(PostEntity postEntity) {
-        PostResponse PostResponse = new PostResponse();
-        PostResponse.setId(postEntity.getId());
-        PostResponse.setContent(postEntity.getContent());
-        PostResponse.setCreatedAt(postEntity.getCreatedAt());
-        PostResponse.setModifiedAt(postEntity.getModifiedAt());
-        PostResponse.setWriter(userController.toResponse(postEntity.getWriter()));
-        return PostResponse;
-    }
-}
