--- Yunbeom/src/main/java/com/depromeet/yunbeom/exception/CertificationCodeNotMatchedException.java
@@ -1,8 +1,0 @@
-package com.depromeet.yunbeom.exception;
-
-public class CertificationCodeNotMatchedException extends RuntimeException {
-
-    public CertificationCodeNotMatchedException() {
-        super("자격 증명에 실패하였습니다.");
-    }
-}
