--- Yunbeom/src/main/java/com/depromeet/yunbeom/model/UserStatus.java
@@ -1,5 +1,0 @@
-package com.depromeet.yunbeom.model;
-
-public enum UserStatus {
-    PENDING, INACTIVE, ACTIVE
-}
