--- Yunbeom/src/main/java/com/depromeet/yunbeom/model/dto/PostResponse.java
@@ -1,15 +1,0 @@
-package com.depromeet.yunbeom.model.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class PostResponse {
-
-    private Long id;
-    private String content;
-    private Long createdAt;
-    private Long modifiedAt;
-    private UserResponse writer;
-}
