--- Yunbeom/src/main/java/com/depromeet/yunbeom/model/dto/UserResponse.java
@@ -1,17 +1,0 @@
-package com.depromeet.yunbeom.model.dto;
-
-import com.depromeet.yunbeom.model.UserStatus;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class UserResponse {
-
-    private Long id;
-    private String email;
-    private String nickname;
-    private UserStatus status;
-    private Long lastLoginAt;
-}
