--- Yunbeom/src/main/java/com/depromeet/yunbeom/model/dto/UserUpdateDto.java
@@ -1,21 +1,0 @@
-package com.depromeet.yunbeom.model.dto;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-public class UserUpdateDto {
-
-    private final String nickname;
-    private final String address;
-
-    @Builder
-    public UserUpdateDto(
-        @JsonProperty("nickname") String nickname,
-        @JsonProperty("address") String address) {
-        this.nickname = nickname;
-        this.address = address;
-    }
-}
