+++ Yunbeom/src/main/java/com/depromeet/yunbeom/post/infrastructure/PostRepository.java
@@ -1,0 +1,7 @@
+package com.depromeet.yunbeom.post.infrastructure;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PostRepository extends JpaRepository<PostEntity, Long> {
+
+}
