--- Yunbeom/src/main/java/com/depromeet/yunbeom/repository/PostRepository.java
@@ -1,7 +1,0 @@
-package com.depromeet.yunbeom.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PostRepository extends JpaRepository<PostEntity, Long> {
-
-}
