+++ Yunbeom/src/main/java/com/depromeet/yunbeom/user/domain/UserStatus.java
@@ -1,0 +1,5 @@
+package com.depromeet.yunbeom.user.domain;
+
+public enum UserStatus {
+    PENDING, INACTIVE, ACTIVE
+}
