+++ Yunbeom/src/main/java/com/depromeet/yunbeom/user/exception/CertificationCodeNotMatchedException.java
@@ -1,0 +1,8 @@
+package com.depromeet.yunbeom.user.exception;
+
+public class CertificationCodeNotMatchedException extends RuntimeException {
+
+    public CertificationCodeNotMatchedException() {
+        super("자격 증명에 실패하였습니다.");
+    }
+}
