+++ Yunbeom/src/main/java/com/depromeet/yunbeom/user/service/UserService.java
@@ -1,0 +1,91 @@
+package com.depromeet.yunbeom.user.service;
+
+import java.time.Clock;
+import java.util.UUID;
+
+import org.springframework.mail.SimpleMailMessage;
+import org.springframework.mail.javamail.JavaMailSender;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import com.depromeet.yunbeom.user.exception.CertificationCodeNotMatchedException;
+import com.depromeet.yunbeom.user.exception.ResourceNotFoundException;
+import com.depromeet.yunbeom.user.domain.UserStatus;
+import com.depromeet.yunbeom.user.domain.UserCreate;
+import com.depromeet.yunbeom.user.domain.UserUpdate;
+import com.depromeet.yunbeom.user.infrastructure.UserEntity;
+import com.depromeet.yunbeom.user.infrastructure.UserRepository;
+
+import lombok.RequiredArgsConstructor;
+
+@Service
+@RequiredArgsConstructor
+public class UserService {
+
+    private final UserRepository userRepository;
+    private final JavaMailSender mailSender;
+
+    @Transactional(readOnly = true)
+    public UserEntity getByEmail(String email) {
+        return userRepository.findByEmailAndStatus(email, UserStatus.ACTIVE)
+            .orElseThrow(() -> new ResourceNotFoundException("Users", email));
+    }
+
+    // get은 애초에 데이터가 없으면 에러를 던진다는 의미가 내포되어 있음
+    // find는 데이터가 없어도 에러를 던지지 않는다는 의미가 내포되어 있음
+    @Transactional(readOnly = true)
+    public UserEntity getById(long id) {
+        return userRepository.findByIdAndStatus(id, UserStatus.ACTIVE)
+            .orElseThrow(() -> new ResourceNotFoundException("Users", id));
+    }
+
+    @Transactional
+    public UserEntity create(UserCreate userCreate) {
+        UserEntity userEntity = new UserEntity();
+        userEntity.setEmail(userCreate.getEmail());
+        userEntity.setNickname(userCreate.getNickname());
+        userEntity.setAddress(userCreate.getAddress());
+        userEntity.setStatus(UserStatus.PENDING);
+        userEntity.setCertificationCode(UUID.randomUUID().toString());
+        userEntity = userRepository.save(userEntity);
+        String certificationUrl = generateCertificationUrl(userEntity);
+        sendCertificationEmail(userCreate.getEmail(), certificationUrl);
+        return userEntity;
+    }
+
+    @Transactional
+    public UserEntity update(long id, UserUpdate userUpdate) {
+        UserEntity userEntity = getById(id);
+        userEntity.setNickname(userUpdate.getNickname());
+        userEntity.setAddress(userUpdate.getAddress());
+        userEntity = userRepository.save(userEntity);
+        return userEntity;
+    }
+
+    @Transactional
+    public void login(long id) {
+        UserEntity userEntity = userRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Users", id));
+        userEntity.setLastLoginAt(Clock.systemUTC().millis());
+    }
+
+    @Transactional
+    public void verifyEmail(long id, String certificationCode) {
+        UserEntity userEntity = userRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Users", id));
+        if (!certificationCode.equals(userEntity.getCertificationCode())) {
+            throw new CertificationCodeNotMatchedException();
+        }
+        userEntity.setStatus(UserStatus.ACTIVE);
+    }
+
+    private void sendCertificationEmail(String email, String certificationUrl) {
+        SimpleMailMessage message = new SimpleMailMessage();
+        message.setTo(email);
+        message.setSubject("Please certify your email address");
+        message.setText("Please click the following link to certify your email address: " + certificationUrl);
+        mailSender.send(message);
+    }
+
+    private String generateCertificationUrl(UserEntity userEntity) {
+        return "http://localhost:8080/api/users/" + userEntity.getId() + "/verify?certificationCode=" + userEntity.getCertificationCode();
+    }
+}
