+++ Yunbeom/src/test/java/com/depromeet/yunbeom/controller/PostCreateControllerTest.java
@@ -1,6 +1,5 @@
 package com.depromeet.yunbeom.controller;
 
-import static org.junit.jupiter.api.Assertions.*;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
@@ -15,7 +14,7 @@ import org.springframework.test.context.jdbc.Sql;
 import org.springframework.test.context.jdbc.SqlGroup;
 import org.springframework.test.web.servlet.MockMvc;
 
-import com.depromeet.yunbeom.model.dto.PostCreateDto;
+import com.depromeet.yunbeom.post.domain.PostCreate;
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 @SpringBootTest
@@ -38,7 +37,7 @@ class PostCreateControllerTest {
 	@Test
 	void 사용자는_게시물을_작성할_수_있다() throws Exception {
 		// given
-		PostCreateDto postCreateDto = PostCreateDto.builder()
+		PostCreate postCreate = PostCreate.builder()
 			.writerId(1)
 			.content("helloworld")
 			.build();
@@ -48,7 +47,7 @@ class PostCreateControllerTest {
 		mockMvc.perform(
 				post("/api/posts")
 					.contentType(MediaType.APPLICATION_JSON)
-					.content(objectMapper.writeValueAsString(postCreateDto)))
+					.content(objectMapper.writeValueAsString(postCreate)))
 			.andExpect(status().isCreated())
 			.andExpect(jsonPath("$.id").isNumber())
 			.andExpect(jsonPath("$.content").value("helloworld"))
